Separated graphing code. Added comments. Fixed Json output bug. Ran clang-format.
