+++ lib/Dialect/HW/CMakeLists.txt
@@ -7,6 +7,9 @@ set(CIRCT_HW_Sources
   HWInstanceGraph.cpp
   HWInstanceImplementation.cpp
   HWModelExplorer.cpp
+  HWJsonInstanceGraph.cpp
+  HWJsonOperationGraph.cpp
+  HWJsonInstanceDiffGraph.cpp
   HWModuleOpInterface.cpp
   HWOpInterfaces.cpp
   HWOps.cpp
