+++ lib/Dialect/HW/Transforms/CMakeLists.txt
@@ -3,7 +3,7 @@ add_circt_dialect_library(CIRCTHWTransforms
   HWPrintInstanceGraph.cpp
   HWSpecialize.cpp
   PrintHWModuleGraph.cpp
-  PrintHWModuleJson.cpp
+  PrintHWModuleOperationJson.cpp
   PrintHWModuleInstanceJson.cpp
   FlattenIO.cpp
   VerifyInnerRefNamespace.cpp
